environment issue

+++ pointnet2/models/clip_encoder.py
@@ -17,7 +17,7 @@ class CLIPEncoder:
             pil_images = [to_pil_image(image) for image in batch_images]
             processed_batch_images = [self.preprocess(pil_image) for pil_image in pil_images]
             batch_images = torch.stack(processed_batch_images).to(self.device)
-            return self.model.encode_image(batch_images)
+            return self.model.encode_image(batch_images).float()
     
     def get_image_features(self, class_index):
         if class_index.ndim > 1:
@@ -28,4 +28,4 @@ class CLIPEncoder:
             processed_text.append(clip.tokenize(self.category_names[index]))
         processed_text = torch.cat(processed_text, dim=0).to(self.device)
         with torch.no_grad():
-            return self.model.encode_text(processed_text)
+            return self.model.encode_text(processed_text).float()
